# Re-write median(Binomial) to improve speed