Fix #1926 by following suggestions therein to reduce avg number of `cdf(Binomial)` evaluations.