@marcusps 2024-12-15
Following the suggestions, given
```julia
rv = Binomial(25,0.3)
rv2 = Binomial(25,0.756)
```
the previous version yielded
```julia
julia> @benchmark median(rv)
BenchmarkTools.Trial: 10000 samples with 33 evaluations.
 Range (min … max):  924.576 ns …   9.896 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     929.455 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   980.001 ns ± 151.139 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▅               ▁     ▁   ▄   ▃                              ▁
  ██▅▇▇▇▇▅▇▅▃▇▄▅▇▁▃█▃██▁▁█▇▅███▃▆█▇▅▃█▆▃▄▇▄▁▁▄█▅▁▃▃█▅▄▁▅▇▆▄▁▁▃▇ █
  925 ns        Histogram: log(frequency) by time       1.49 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark median(rv2)
BenchmarkTools.Trial: 10000 samples with 26 evaluations.
 Range (min … max):  941.346 ns …  13.187 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     945.231 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.007 μs ± 187.974 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █               ▁   ▂  ▂   ▄   ▃                              ▁
  █▇█▅▅█▅▆▇▄█▅▁▇▅▃██▁▃█▃▁█▅▁▃█▃▁▄█▅▆▇█▄▅▅█▇▄▃▁█▄▁▃▃█▃▁▁▁▆█▁▄▄▁█ █
  941 ns        Histogram: log(frequency) by time       1.51 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
and the version version yields
```julia
julia> @benchmark median(rv)
BenchmarkTools.Trial: 10000 samples with 469 evaluations.
 Range (min … max):  225.928 ns … 557.132 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     228.358 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   235.036 ns ±  21.306 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ██▂▆▇▁                          ▁▁   ▁                        ▂
  ████████▅▅▅▇▇▆▇▅█▇▄▄█▆▄▅▆▅▄▄▅▅▄▄██▆▅▄█▇▅▅▄▇▅▃▄▂▄▄▃▄▄▄▆▄▃▄▄▅▅▄ █
  226 ns        Histogram: log(frequency) by time        340 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark median(rv2)
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  17.037 ns … 52.826 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     18.434 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.582 ns ±  2.704 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▅██▇▇▅▆▆▅▃     ▁  ▁▂  ▁▁▃▂▄▄▄▃▃▂▃▂ ▁▁                     ▂
  ▄▇██████████▇█▇▄███▇██▇▇███████████████▇▆▇▇▇▆▇▇▇▄▆▇▃▇▆▅█▂▇▆ █
  17 ns        Histogram: log(frequency) by time        29 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
I will check how much the  `minmax` suggestion helps.