@marcusps 2024-12-16
Here is a more thorough comparison.  To get full coverage across the branches, I extended the benchmarks to
```julia
rv = Binomial(25,0.3)
rv2 = Binomial(25,0.756)
rv3 = Binomial(25,1//2)
rv4 = Binomial(25, 3//5)
```
and then ran `bitj = @benchmark median(rvj)` where `i`=1 represents the current (buggy) implementation of the median, `i`=2 is the current (merged) implementation that fixes the bug, and `i=3` incorporates something akin to `minmax` (but still using only `min`), which looks like this:
```julia
function median(dist::Binomial)
    bound = min(dist.p, 1-dist.p) # from http://dx.doi.org/10.1111/j.1467-9574.1980.tb00681.x
    dist_mean = mean(dist)
    
    floor_mean = floor(Int, dist_mean)
    difference = dist_mean - floor_mean
    
    if difference < bound
        floor_mean
    elseif  difference > 1 - bound
        floor_mean + 1
    elseif cdf(dist, floor_mean) >= 0.5
        floor_mean
    else
        floor_mean + 1
    end
end
```
The results are
```julia
julia> for (a,b) in zip(b2,b3)
         println(BenchmarkTools.judge(median(a),median(b)))
       end
TrialJudgement(+0.40% => invariant)
TrialJudgement(-9.69% => improvement)
TrialJudgement(-0.04% => invariant)
TrialJudgement(-3.12% => invariant)

julia> for (a,b) in zip(b2,b3)
         println(BenchmarkTools.judge(var(a),var(b)))
       end
TrialJudgement(+12.14% => regression)
TrialJudgement(-49.32% => improvement)
TrialJudgement(-22.95% => improvement)
TrialJudgement(-12.07% => improvement)
```
which means the simplification is a definite improvement, both for the median and variance.

The new bug free implementations are still much slower than the buggy implementation, but that feels like an unfair comparison.
```julia
julia> for (a,b) in zip(b1,b2)
         println(BenchmarkTools.judge(median(a),median(b)))
       end
TrialJudgement(+308.74% => regression)
TrialJudgement(+5373.54% => regression)
TrialJudgement(+343.01% => regression)
TrialJudgement(+3152.94% => regression)

julia> for (a,b) in zip(b1,b3)
         println(BenchmarkTools.judge(median(a),median(b)))
       end
TrialJudgement(+310.37% => regression)
TrialJudgement(+4843.23% => regression)
TrialJudgement(+342.83% => regression)
TrialJudgement(+3051.34% => regression)
```