@devmotion 2024-12-16
I see the following difference to the master branch:

### This PR

```julia
julia> using Distributions, Chairmarks

julia> @be median($(Binomial(44, 0.1)))
Benchmark: 3317 samples with 186 evaluations
 min    135.301 ns
 median 152.554 ns
 mean   153.703 ns
 max    250.892 ns

julia> @be median($(Binomial(40, 0.1)))
Benchmark: 4932 samples with 4265 evaluations
 min    3.976 ns
 median 4.298 ns
 mean   4.556 ns
 max    9.066 ns

julia> @be median($(Binomial(25, 0.3)))
Benchmark: 3891 samples with 147 evaluations
 min    137.755 ns
 median 157.592 ns
 mean   163.907 ns
 max    284.014 ns

julia> @be median($(Binomial(25, 0.756)))
Benchmark: 4933 samples with 3917 evaluations
 min    4.266 ns
 median 4.606 ns
 mean   4.900 ns
 max    21.966 ns

julia> @be median($(Binomial(25, 0.5)))
Benchmark: 4809 samples with 4545 evaluations
 min    3.988 ns
 median 4.171 ns
 mean   4.351 ns
 max    9.332 ns

julia> @be median($(Binomial(25, 0.6)))
Benchmark: 4921 samples with 4227 evaluations
 min    3.982 ns
 median 4.298 ns
 mean   4.538 ns
 max    9.148 ns
```

### Master branch

```julia
julia> using Distributions, Chairmarks

julia> @be median($(Binomial(44, 0.1)))
Benchmark: 3387 samples with 125 evaluations
 min    214.664 ns
 median 215.672 ns
 mean   222.091 ns
 max    366.672 ns

julia> @be median($(Binomial(40, 0.1)))
Benchmark: 4111 samples with 86 evaluations
 min    239.826 ns
 median 253.395 ns
 mean   265.475 ns
 max    431.686 ns

julia> @be median($(Binomial(25, 0.3)))
Benchmark: 3938 samples with 101 evaluations
 min    215.347 ns
 median 225.663 ns
 mean   234.931 ns
 max    417.901 ns

julia> @be median($(Binomial(25, 0.756)))
Benchmark: 3951 samples with 94 evaluations
 min    225.617 ns
 median 242.904 ns
 mean   250.102 ns
 max    419.777 ns

julia> @be median($(Binomial(25, 0.5)))
Benchmark: 3920 samples with 100 evaluations
 min    216.250 ns
 median 227.080 ns
 mean   237.701 ns
 max    419.170 ns

julia> @be median($(Binomial(25, 0.6)))
Benchmark: 3643 samples with 105 evaluations
 min    217.057 ns
 median 233.733 ns
 mean   245.306 ns
 max    732.143 ns
```